+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - the Lagrange-interpolation grid of PineAPPL now supports fully dynamic
   factorisation/renormalisation scales
+- `pineappl` has a new subcommand `pdf_uncertainty` to calculate PDF
+  uncertainties
 
 ## [0.1.0] - 11/06/2020